+++ Src/MoneyManager/MoneyManager.WindowsPhone/Common/NavigationHelper.cs
@@ -1,14 +1,10 @@
-﻿#region
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using Windows.Foundation.Metadata;
 using Windows.Phone.UI.Input;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Navigation;
-
-#endregion
 
 namespace MoneyManager.Common
 {
