+++ Src/MoneyManager/MoneyManager.WindowsPhone/Src/TileHelper.cs
