+++ Src/MoneyManager/MoneyManager.WindowsPhone/Src/UserNotification.cs
