+++ Src/MoneyManager/MoneyManager.WindowsPhone/Src/ViewModelLocator.cs
