+++ Src/MoneyManager/MoneyManager.WindowsPhone/Views/MainPage.xaml
@@ -1,5 +1,5 @@
 ﻿<Page
-    x:Class="MoneyManager.MainPage"
+    x:Class="MoneyManager.Views.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
